% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_id.R
\name{mask_id}
\alias{mask_id}
\title{Mask an Input String}
\usage{
mask_id(string, seed = 395)
}
\arguments{
\item{string}{A character string to be masked.}

\item{seed}{A single value, interpreted as an integer, to set the seed of
R's random number generator for reproducibility. Default is 395.}
}
\value{
An MD5 hashed representation of the masked string.
}
\description{
This function masks a given input string to protect sensitive data such as IDs.
It works by shuffling the input string characters and then computing its MD5 hash.
This method ensures a consistent hash for a given string when the same seed is used,
allowing for reproducibility while ensuring privacy.
}
\examples{
mask("sampleID")
mask("sampleID", seed = 123)

}
\seealso{
\href{https://researchbox.org/messages.php?id=replace_private_id}{Masking IDs in ResearchBox}
}
