% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_from_qsf.R
\name{read_csv_from_qsf}
\alias{read_csv_from_qsf}
\title{Read Qualtrics CSV Files and Clean Column Names}
\usage{
read_csv_from_qsf(file_location)
}
\arguments{
\item{file_location}{A string indicating the path to the CSV file to be read.}
}
\value{
A data frame with cleaned column names and without Qualtrics metadata.
}
\description{
This function reads a CSV file exported from Qualtrics, extracts the variable names, 
skips the metadata, and returns a cleaned data frame with standardized column names.
}
\details{
The function performs the following steps:
1. Reads the CSV file without assuming any column names.
2. Extracts the variable names from the first row of the read data.
3. Re-reads the data, skipping the first three rows (Qualtrics metadata) and uses the extracted variable names.
4. Cleans the variable names using the `janitor::clean_names` function.
}
\examples{
# Ensure you have set your working directory or provide an absolute path
data <- read_csv_from_qsf("your_qualtrics_file.csv")

}
