% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_code.R
\name{reverse_code}
\alias{reverse_code}
\title{Reverse code a variable based on specified original range}
\usage{
reverse_code(variable, min_orig, max_orig)
}
\arguments{
\item{variable}{A numeric vector to be reverse coded.}

\item{min_orig}{The original minimum value of the range.}

\item{max_orig}{The original maximum value of the range.}
}
\value{
A numeric vector of the same length as `variable`, with values reverse coded.
}
\description{
This function reverse codes a variable so that the original minimum becomes the maximum and vice versa.
It checks to ensure that all values in the variable are within the specified original range.
}
\examples{
reverse_code(c(1, 2, 3, 4, 5), 1, 5)
reverse_code(c(-2, -1, 0, 1, 2), -2, 2)

}
