% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_select_all_responses.R
\name{summarize_select_all_responses}
\alias{summarize_select_all_responses}
\title{Summarize 'Select All' Survey Responses}
\usage{
summarize_select_all_responses(.data, prefix, renaming_scheme = NULL)
}
\arguments{
\item{.data}{A dataframe containing the survey data.}

\item{prefix}{The common prefix of the 'Select All' indicator variables in the dataset.}

\item{renaming_scheme}{(Optional) A named list or vector for renaming the extracted categories. If not provided, categories won't be renamed.}
}
\value{
A dataframe with columns for the response option, proportion of selections, and standard error of the proportions.
}
\description{
This function compiles and summarizes indicator variables from survey data where participants could "Select All" responses that applied. It then computes the proportion of selections and their standard errors for each response option.
}
\examples{
\dontrun{
data_clean <- data.frame(
  select_all_type_of_conv_describeIdea = sample(c(0,1), 100, replace = TRUE),
  select_all_type_of_conv_describeEmotion = sample(c(0,1), 100, replace = TRUE)
)
summarized_data <- summarize_select_all_responses(data_clean, "select_all_type_of_conv")
print(summarized_data)
}

}
